Improved kbd driver and input device to use keycodes

+++ sys/drivers/platform/kbd/kbd.h
@@ -10,7 +10,7 @@ struct input_device;
 
 struct kbd_private {
     bool caps_lock;
-    uint8_t *scancode_set;
+    unsigned int *scancode_set;
     struct input_device *idev;
 };
